Move `ExternalPool` to `components` folder
